Fix linting, tests and route generation
