+++ apps/connect-api/package.json
@@ -3,8 +3,10 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "start": "npx tsx --watch ./src/main.ts",
+    "mount-routes": "npx tsx ./src/scripts/generateRoutes.ts && npm run lint",
+    "start": "npm run mount-routes && npx tsx --watch ./src/main.ts",
     "start:prod": "npx tsx ./dist/apps/connect-api/src/main.js",
+    "lint": "npx eslint ./src --fix -c  ./.eslintrc.json",
     "build": "rm -rf ./dist && ../../node_modules/typescript/bin/tsc --project ./tsconfig.json",
     "test": "npx jest --config=./jest.config.ts",
     "echo": "echo \"Builder api\""
