+++ apps/connect-api/src/api/random-number.ts
@@ -1,5 +1,4 @@
 import { PathBasedRouter } from 'connect-api/lib/pathBasedRouter';
-
 import { randomIntFromInterval } from 'lib/utils/random';
 
 const router = new PathBasedRouter();
