+++ apps/connect-api/tsconfig.json
@@ -14,6 +14,6 @@
     "strict": true,                                      /* Enable all strict type-checking options. */
     "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
   },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules"],
+  "include": ["src/**/*.ts", "./jest.config.ts"],
+  "exclude": ["node_modules", "dist"],
 }
